Fixed user registration process requiring address
Excluded address fields from UserIn_Pydantic model so that those fields
are not mandatory for account creation. seperate endpoint should be
implemented to add address. Work booing should not be allowed if address
fields are not added by the user.

upgrading the user account to a worker account should require adding
address. which can use the same API for adding the address